move cache and base64utils to pkg so they can be imported into verification server codebase (#628)
